spawn/NamespaceOptions: map root user+group into new user namespaces
If the process is set up by the root superuser but its uid/gid is not
mapped, then mkdir() on a new tmpfs will fail with EOVERFLOW, and thus
MOUNT_ROOT_TMPFS is doomed to always fail, because it cannot create
mountpoints.

One approach would be to map root only if MOUNT_ROOT_TMPFS is enabled,
but on the other hand, always mapping the root user+group doesn't
hurt, and may clarify system file ownerships.