spawn/Launch: failure to create systemd scope is fatal
Recently, we had failures to create the scope, e.g.:

 "Failed to create systemd scope: Did not receive a reply. Possible
 causes include: the remote application did not send a reply, the
 message bus security policy blocked the reply, the reply timeout
 expire"

Because this error was non-fatal, the process started successfully,
but was unable to do actual work.  Since we don't have a server where
cgroups are not used, let's make this feature mandatory (with a
configuration setting to make it optional again).