spawn/Direct: check UidGid::IsNop() in parent process
This fixes breakages in "debug mode" where the application runs on a
developer's machine as regular user and can never switch users.  This
check needs to be done before the new user namespace is created, or
else getresuid() will return the "overflow" uid and IsNop() inside
that namespace will always return false.