co/MultiAwaitable: implement a custom task to handle deletion properly
Previously, the coroutine running Wait() could be deleted while it was
running if the MultiAwaitable containing the EagerTask was deleted.
Coroutine handles, however, must not be deleted while they are
running, therefore we must postpone the deletion until after it was
finished.  By implementing a custom task, we can let the compiler
generate code for this after final_suspend().