stock/MultiStock: consider unclean items for the concurrency limit
The concurrency was only compared with the number of busy items, but
unclean idle items are just as unavailable.  Therefore, if no clean
idle item is available, create a new lease only if the total of all
items (idle and busy) is still below the concurrency limit.
Previously, this was unbounded if there were lots of unclean items.