stock/MultiStock: do not discard idle items if there are waiters
If there are waiters, this means the items have just become available
but have not yet been submitted to those waiters (via `retry_event`).
Or they are "unclean" and cannot yet be submitted.

This fixes a stall bug which can occur if unclean items have been
discarded which means that ItemUncleanFlagCleared() never got called
and waiters will never be called back.