thread/Worker: configure 128 kB stack on ARM64
The pthread_attr_setstacksize() call fails on ARM64 because 64 kB is
smaller than PTHREAD_STACK_MIN.  glibc then ignores the call
completely and allocates 8 MB stack size.  This is surprising because
pthread_attr_setstacksize() is only supposed to set the minimum stack
size, and it seems more reasonable to simply use PTHREAD_STACK_MIN if
the parameter is smaller than that.  Duh!