translation/Parser: support per-SERVICE ExecuteOptions
The SERVICE packet is now allowed in translation responses and it
begins a new ExecuteOptions section for that service.  That way, only
one translation request is needed for a SSH connection; if the SSH
clients requests a SFTP channel, the SERVICE=sftp part from the
initial translation response can simply be used instead of sending
another translation request.