Revert "event/Loop: don't repeat epoll_wait() if all events were received"
This reverts commit d9fbba5fcbf520b60552db4bcdcbbad64eca4f62.  It was
based on a fallacy.  It is true that io_uring poll is edge-triggered,
but my assumption that a non-full epoll_wait() makes the epoll
descriptor "not ready".  It remains ready as long as any registered
file descriptor is ready, and as long as it's ready, we should be
reading and handling further events from it.  The commit led to
stalled transfers because we did not handle those (repeated) events.